projects/app-web-django-assignment/requirements.txt
django
django-filter
django-allauth
django-apscheduler
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
